completed output
output is now successfully done via information stored in UserMarker.class instead of via a new object (that was dumb, and I shouldn't have done it that way in the first place), and version has been updated to 0.3 with the alpha prefix removed. I also changed the way the on-screen display for coordinates is updated, and it should be more accurate now.