updated pre-made auton graphing abilities
autonCalc is now capable of isolating the number values given as arguments to commands that move the bot in a coordinate plane (basically, if it moves the bot in a way that changes its position and not, for instance, raises the elevator, because as of the moment there are no plans to graph such things). Next step is to add the most challenging piece - actually turning those values into drawable UserMarker objects, then hook up the gui to the sequencerReader object.