updates to auto graphing
more complete parsing and the beginnings of the marker building have been implemented, and the graphical front end has been hooked up (albeit in a kind of rudimental way) to the parsing engine. Going forward, cleaning up the spaghetti code in board, redoing some parts of the graphics in board, and finishing the marker synthesis portion of the SequencerReader class are the immediate goals.